Split geometry utilities and parametric curve classes into two files
